requirements.txt
ansible==2.1.1.0
cffi==1.7.0
cryptography==1.5
docopt==0.6.2
enum34==1.1.6
idna==2.1
ipaddress==1.0.16
jedi==0.9.0
Jinja2==2.8
MarkupSafe==0.23
paramiko==2.0.2
prompt-toolkit==1.0.5
ptpython==0.35
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
Pygments==2.1.3
pyvmomi==6.0.0.2016.6
PyYAML==3.12
requests==2.11.0
scapy==2.3.2
six==1.10.0
wcwidth==0.1.7
